requirements.txt
aiohttp==3.8.1
aiosignal==1.3.2
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.2
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
colorama==0.4.6
cssselect==1.2.0
exceptiongroup==1.1.3
fake-useragent==1.5.1
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.5.2
lxml==4.9.3
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
platformdirs==3.10.0
propcache==0.2.1
pycparser==2.22
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-cache==1.0.0
requests-html==0.10.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.29
tqdm==4.66.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.1
websockets==10.4
wsproto==1.2.0
yarl==1.18.3
zipp==3.18.1
